CV/requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git

# API server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.0

# Image processing
Pillow>=10.0.0

# Utilities
numpy>=1.24.0
python-dotenv>=1.0.0

# Deployment
modal>=0.60.0
requests>=2.31.0